refactor: 拆分超限适配器与测试模块
