+++ tests/test_responses_provider.py
@@ -21,7 +21,6 @@ from opencollab.adapters.llm.responses_provider import (
     _messages_to_input,
     _parse_stream,
     complete_responses,
-    parse_responses_response,
 )
 
 
@@ -616,21 +615,25 @@ def test_item_replay_binds_function_call_output_to_original_call_id():
 
 
 def test_responses_preserves_image_url_content_blocks():
-    messages = [{
-        "role": "user",
-        "content": [
-            {"type": "text", "text": "What is shown?"},
-            {"type": "image_url", "image_url": {"url": "https://example.test/a.png", "detail": "high"}},
-        ],
-    }]
+    messages = [
+        {
+            "role": "user",
+            "content": [
+                {"type": "text", "text": "What is shown?"},
+                {"type": "image_url", "image_url": {"url": "https://example.test/a.png", "detail": "high"}},
+            ],
+        }
+    ]
     _, items = _messages_to_input(messages)
-    assert items == [{
-        "role": "user",
-        "content": [
-            {"type": "input_text", "text": "What is shown?"},
-            {"type": "input_image", "image_url": "https://example.test/a.png", "detail": "high"},
-        ],
-    }]
+    assert items == [
+        {
+            "role": "user",
+            "content": [
+                {"type": "input_text", "text": "What is shown?"},
+                {"type": "input_image", "image_url": "https://example.test/a.png", "detail": "high"},
+            ],
+        }
+    ]
 
 
 def test_responses_rejects_tools_for_models_without_function_calling():
@@ -755,73 +758,3 @@ def test_prompt_cache_key_is_stable_within_one_run_and_isolated_between_runs():
     assert first["prompt_cache_key"] != other_run["prompt_cache_key"]
     assert first["prompt_cache_key"] != other_client["prompt_cache_key"]
     assert len(first["prompt_cache_key"]) == 64
-
-
-def test_non_streaming_text_and_missing_usage_are_supported():
-    response = completed_response(output=[message_item("done")])
-    parsed = parse_responses_response(
-        response,
-        [{"role": "user", "content": "work"}],
-        expected_model="gpt-fake",
-    )
-    assert parsed.content == "done"
-    assert parsed.usage.estimated is False
-
-    response.usage = None
-    estimated = parse_responses_response(
-        response,
-        [{"role": "user", "content": "work"}],
-        expected_model="gpt-fake",
-    )
-    assert estimated.usage.estimated is True
-    assert estimated.usage.cache_read_tokens is None
-    assert estimated.usage.reasoning_tokens is None
-
-
-def test_non_streaming_usage_accepts_top_level_cache_write_fallback():
-    response = completed_response(output=[message_item("done")])
-    del response.usage.input_tokens_details.cache_write_tokens
-    response.usage.cache_write_tokens = 4
-
-    parsed = parse_responses_response(
-        response,
-        [{"role": "user", "content": "work"}],
-        expected_model="gpt-fake",
-    )
-
-    assert parsed.usage.cache_creation_tokens == 4
-
-
-def test_non_streaming_requires_provider_model_identity():
-    response = completed_response(output=[message_item("done")], model="")
-
-    with pytest.raises(ResponsesProtocolError, match="missing model identity"):
-        parse_responses_response(
-            response,
-            [{"role": "user", "content": "work"}],
-            expected_model="gpt-fake",
-        )
-
-
-def test_non_streaming_records_provider_resolved_model_alias_and_estimates_invalid_usage():
-    aliased = completed_response(output=[message_item("done")], model="other-model")
-    parsed_alias = parse_responses_response(
-        aliased,
-        [{"role": "user", "content": "work"}],
-        expected_model="gpt-fake",
-    )
-    assert parsed_alias.provider_model == "other-model"
-
-    response = completed_response(output=[message_item("done")])
-    response.usage.input_tokens = -1
-    response.usage.output_tokens = 0
-    response.usage.input_tokens_details.cached_tokens = -4
-    parsed = parse_responses_response(
-        response,
-        [{"role": "user", "content": "work"}],
-        expected_model="gpt-fake",
-    )
-    assert parsed.usage.estimated is True
-    assert parsed.usage.input_tokens > 0
-    assert parsed.usage.output_tokens > 0
-    assert parsed.usage.cache_read_tokens is None
